fix: layout.tsx를 서버 컴포넌트로 수정
